set up svgoconfig properly in supportPath.

+++ extensions/svgr/src/index.tsx
@@ -1,18 +1,21 @@
+import { existsSync, writeFileSync } from "fs";
+import { useEffect } from "react";
 import {
-  Form,
-  ActionPanel,
   Action,
+  ActionPanel,
   Clipboard,
+  Form,
+  Icon,
   LocalStorage,
+  closeMainWindow,
+  confirmAlert,
+  environment,
   showHUD,
   showToast,
-  closeMainWindow,
-  Icon,
 } from "@raycast/api";
-import { useEffect } from "react";
 import Settings from "./components/Settings";
 import { getReactSVG } from "./util";
-import { svgrDefaultSettings } from "./constants";
+import { svgrDefaultSettings, svgoDefaultSettings } from "./constants";
 
 export interface Props {
   componentName: string;
@@ -20,6 +23,8 @@ export interface Props {
 }
 
 export default function Command() {
+  const svgoConfigPath = `${environment.supportPath}/.svgorc.json`;
+
   useEffect(() => {
     (async () => {
       const localSvgrSettings = await LocalStorage.getItem("svgr");
@@ -27,24 +32,45 @@ export default function Command() {
         await LocalStorage.setItem("svgr", JSON.stringify(svgrDefaultSettings));
       }
     })();
+    const svgoSettings = existsSync(svgoConfigPath);
+    if (!svgoSettings) writeFileSync(svgoConfigPath, JSON.stringify(svgoDefaultSettings));
   }, []);
 
   const handleCopy = async (props: Props) => {
-    const reactSVG = await getReactSVG(props);
+    const reactSVG = await getReactSVG({ ...props, svgoConfigPath });
     await Clipboard.copy(reactSVG);
     await showHUD("Copied to Clipboard");
   };
 
   const handlePaste = async (props: Props) => {
-    const reactSVG = await getReactSVG(props);
+    const reactSVG = await getReactSVG({ ...props, svgoConfigPath });
     await Clipboard.paste(reactSVG);
     await closeMainWindow();
   };
 
-  const handleClearLocalStorage = async () => {
-    await LocalStorage.removeItem("svgr");
-    await LocalStorage.setItem("svgr", JSON.stringify(svgrDefaultSettings));
-    await showToast({ title: "Restore Default Settings", message: "Success! Default Settings Restored." });
+  const handleRestoreSvgrSettings = async () => {
+    const confirmed = await confirmAlert({
+      title: "Restore Default SVGR Settings?",
+      message: "You cannot undo this action.",
+      icon: Icon.Trash,
+    });
+    if (confirmed) {
+      await LocalStorage.removeItem("svgr");
+      await LocalStorage.setItem("svgr", JSON.stringify(svgrDefaultSettings));
+      await showToast({ title: "Restore Default Settings", message: "Success!" });
+    }
+  };
+
+  const handleRestoreSvgoSettings = async () => {
+    const confirmed = await confirmAlert({
+      title: "Restore Default SVGO Settings?",
+      message: "You cannot undo this action.",
+      icon: Icon.Trash,
+    });
+    if (confirmed) {
+      writeFileSync(svgoConfigPath, JSON.stringify(svgoDefaultSettings));
+      await showToast({ title: "Restore Default Settings", message: "Success!" });
+    }
   };
 
   return (
@@ -54,7 +80,17 @@ export default function Command() {
           <Action.SubmitForm icon={Icon.Clipboard} title="Copy to Clipboard" onSubmit={handleCopy} />
           <Action.SubmitForm icon={Icon.Clipboard} title="Paste in Active App" onSubmit={handlePaste} />
           <Action.Push icon={Icon.Gear} title="Customize SVGR Settings" target={<Settings />} />
-          <Action.SubmitForm icon={Icon.Trash} title="Restore Default Settings" onSubmit={handleClearLocalStorage} />
+          <Action.Open icon={Icon.Document} title="Open SVGO Settings" target={svgoConfigPath} />
+          <Action.SubmitForm
+            icon={Icon.TwoArrowsClockwise}
+            title="Restore Default SVGR Settings"
+            onSubmit={handleRestoreSvgrSettings}
+          />
+          <Action.SubmitForm
+            icon={Icon.TwoArrowsClockwise}
+            title="Restore Default SVGO Settings"
+            onSubmit={handleRestoreSvgoSettings}
+          />
         </ActionPanel>
       }
     >
